fix: dataclass 분할및 최상단 이동 기능
