+++ app/src/main/java/com/example/miniyoutube/data/model/remote/searchvideo/Id.kt
@@ -1,0 +1,14 @@
+package com.example.miniyoutube.data.model.remote.searchvideo
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.Parcelize
+
+
+@Parcelize
+data class Id(
+    @SerializedName("kind")
+    val kind: String,
+    @SerializedName("videoId")
+    val videoId: String
+) : Parcelable
