+++ app/src/main/java/com/example/miniyoutube/data/model/remote/searchvideo/Item.kt
@@ -1,0 +1,16 @@
+package com.example.miniyoutube.data.model.remote.searchvideo
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.Parcelize
+
+
+@Parcelize
+data class Item(
+    @SerializedName("id")
+    val id: Id,
+    @SerializedName("snippet")
+    val snippet: Snippet,
+    @SerializedName("contentDetails")
+    val contentDetails: ContentDetails
+) : Parcelable
