+++ app/src/main/java/com/example/miniyoutube/data/model/remote/searchvideo/Medium.kt
@@ -1,0 +1,16 @@
+package com.example.miniyoutube.data.model.remote.searchvideo
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.Parcelize
+
+
+@Parcelize
+data class Medium(
+    @SerializedName("url")
+    val url: String,
+    @SerializedName("width")
+    val width: Int,
+    @SerializedName("height")
+    val height: Int
+) : Parcelable
