+++ app/src/main/java/com/example/miniyoutube/data/model/remote/searchvideo/Thumbnails.kt
@@ -1,0 +1,12 @@
+package com.example.miniyoutube.data.model.remote.searchvideo
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.Parcelize
+
+
+@Parcelize
+data class Thumbnails(
+    @SerializedName("medium")
+    val medium: Medium
+) : Parcelable
