+++ app/src/main/java/com/example/miniyoutube/data/model/remote/searchvideo/YoutubeResponse.kt
@@ -1,0 +1,15 @@
+package com.example.miniyoutube.data.model.remote.searchvideo
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class YoutubeVideo(
+    @SerializedName("kind")
+    val kind: String,
+    @SerializedName("etag")
+    val etag: String,
+    @SerializedName("items")
+    val items: List<Item>
+) : Parcelable
