+++ app/src/main/java/com/example/miniyoutube/data/model/remote/videoinfo/Statistics.kt
@@ -1,0 +1,12 @@
+package com.example.miniyoutube.data.model.remote.videoinfo
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.Parcelize
+
+
+@Parcelize
+data class Statistics(
+    @SerializedName("viewCount")
+    val viewCount: String? = ""
+) : Parcelable
