+++ app/src/main/java/com/example/miniyoutube/data/model/remote/videoinfo/YoutubeVideoInfo.kt
@@ -1,0 +1,15 @@
+package com.example.miniyoutube.data.model.remote.videoinfo
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class YoutubeVideoInfo(
+    @SerializedName("kind")
+    val kind: String,
+    @SerializedName("etag")
+    val etag: String,
+    @SerializedName("items")
+    val items: List<TrendItem>?
+) : Parcelable
