+++ app/src/main/java/com/example/miniyoutube/data/repository/YoutubeRepository.kt
@@ -1,7 +1,7 @@
 package com.example.miniyoutube.data.repository
 
-import com.example.miniyoutube.data.model.remote.YoutubeVideo
-import com.example.miniyoutube.data.model.remote.YoutubeVideoInfo
+import com.example.miniyoutube.data.model.remote.searchvideo.YoutubeVideo
+import com.example.miniyoutube.data.model.remote.videoinfo.YoutubeVideoInfo
 
 interface YoutubeRepository {
     suspend fun requestSearch(q: String, videoCategoryId : String) : YoutubeVideo
