+++ app/src/main/java/com/example/miniyoutube/ui/home/HomeViewModel.kt
@@ -4,13 +4,10 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.miniyoutube.data.model.remote.YoutubeVideo
-import com.example.miniyoutube.data.model.remote.YoutubeVideoInfo
+import com.example.miniyoutube.data.model.remote.videoinfo.YoutubeVideoInfo
 import com.example.miniyoutube.data.repository.YoutubeRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 @HiltViewModel
