+++ app/src/main/java/com/example/miniyoutube/ui/home/recyclerview/CategoryRecyclerViewAdapter.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.example.miniyoutube.data.model.remote.TrendItem
+import com.example.miniyoutube.data.model.remote.videoinfo.TrendItem
 import com.example.miniyoutube.databinding.ItemListVarBinding
 
 class CategoryRecyclerViewAdapter(
