+++ app/src/main/java/com/example/miniyoutube/ui/search/SearchViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.miniyoutube.data.model.remote.YoutubeVideo
+import com.example.miniyoutube.data.model.remote.searchvideo.YoutubeVideo
 import com.example.miniyoutube.data.repository.YoutubeRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
