+++ app/src/main/java/com/example/miniyoutube/ui/search/recyclerview/SearchAdapter.kt
@@ -6,8 +6,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.example.miniyoutube.data.model.remote.Item
-import com.example.miniyoutube.data.model.remote.Snippet
+import com.example.miniyoutube.data.model.remote.searchvideo.Item
 import com.example.miniyoutube.databinding.ViewholderSearchBinding
 
 class SearchAdapter(val onClick: (Item) -> Unit): ListAdapter<Item, SearchAdapter.ViewHolder>(diffUtil) {
